requirements.txt
fastapi
uvicorn
pydantic>=2
jinja2
pyyaml
openai>=1
